Moved MungeTable and PasswordChain to resources package
